Split out BearerToken into seperate file
This commit moves the `BearerToken` enum out of mod.rs and into its own
file to help keep mod.rs focused.
